updated make_image
